Added registry module types with catalog type with internal ids, ObjectStore's use now DataView's istead of DataContainer's and cleanup.

+++ src/main/java/org/lanternpowered/server/network/vanilla/message/codec/play/CodecPlayOutPlayerJoinGame.java
@@ -41,7 +41,7 @@ public final class CodecPlayOutPlayerJoinGame implements Codec<MessagePlayOutPla
         context.getChannel().attr(PLAYER_ENTITY_ID).set(message.getEntityId());
         ByteBuffer buf = context.byteBufAlloc().buffer();
         buf.writeInteger(message.getEntityId());
-        byte gameMode = message.getGameMode().getInternalId();
+        byte gameMode = (byte) message.getGameMode().getInternalId();
         if (message.isHardcore()) {
             gameMode |= 0x8;
         }
